requirements.txt
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.1
django-admin-logs==1.3.0
django-allauth==0.61.1
django-cors-headers==4.4.0
django-daisy==1.0.9
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
iniconfig==2.0.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pika==1.3.2
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.2
